+++ changelog.md
@@ -1,5 +1,5 @@
 - Profiling events are no longer emitted unless a performance profile is being gathered.
-- Changed profiler shortcut to CTRL + ALT + F8.
+- Changed profiler shortcut to CTRL + SHIFT + F8.
 - Added support for javaagent. Suggested by President Matt Damon.
 
 v0.6.8